Backend: Refactor models to shared package and fix error handling
- Extract User and FeatureFlag structs into internal/models package
- Update main.go and seed.go to import from shared models package
- Fix unhandled error in updateFeatureFlagHandler when reloading flag
- Prevents cache inconsistency if database reload fails after update
